Separate models module
Enables splitting things that depend on models, like plotting and
running the sim, so that overrides for dataframe dependencies don't play
poorly with other deps from e.g. lens
